requirements.txt
gradio==5.37.0
langchain==0.3.26
langchain_core==0.3.68
pdfplumber==0.11.7
Pillow==11.3.0
pydantic==2.11.7
pytesseract==0.3.13
python-dotenv==1.1.1
Requests==2.32.4
